aider: Removed the `install` script from `package.json` to prevent an infinite loop during installation.
# Aider chat conversation:

USER: when i try to run this code i get stuck in a loop on render: > cms-style-blog@1.0.0 install
> npm install
> cms-style-blog@1.0.0 install
> npm install
> cms-style-blog@1.0.0 install
> npm install

ASSISTANT: The issue is that the `install` script in `package.json` is causing an infinite loop by calling `npm install` repeatedly. We need to remove the `install` script from `package.json`.

Here is the *SEARCH/REPLACE* block for `package.json`:

package.json
```json
<<<<<<< SEARCH
  "scripts": {
    "install": "npm install",
    "start": "node server.js",
    "dev": "nodemon server.js"
  },
=======
  "scripts": {
    "start": "node server.js",
    "dev": "nodemon server.js"
  },
>>>>>>> REPLACE
```